chore: seperate to seperate classes to make it cleaner
