+++ src/main/java/net/unjoinable/skyblock/event/listener/player/PlayerBlockInteractListener.java
@@ -1,0 +1,18 @@
+package net.unjoinable.skyblock.event.listener.player;
+
+import net.minestom.server.event.player.PlayerBlockInteractEvent;
+
+import java.util.function.Consumer;
+
+import static net.unjoinable.skyblock.event.listener.player.PlayerListenerConstants.IGNORE_ANIMATION;
+
+/**
+ * Handles player block interaction events.
+ */
+public class PlayerBlockInteractListener implements Consumer<PlayerBlockInteractEvent> {
+
+    @Override
+    public void accept(PlayerBlockInteractEvent event) {
+        event.getPlayer().setTag(IGNORE_ANIMATION, true);
+    }
+}
