+++ src/main/java/net/unjoinable/skyblock/event/listener/player/PlayerFinishDiggingListener.java
@@ -1,0 +1,18 @@
+package net.unjoinable.skyblock.event.listener.player;
+
+import net.minestom.server.event.player.PlayerFinishDiggingEvent;
+
+import java.util.function.Consumer;
+
+import static net.unjoinable.skyblock.event.listener.player.PlayerListenerConstants.IS_DIGGING;
+
+/**
+ * Handles player finish digging events.
+ */
+public class PlayerFinishDiggingListener implements Consumer<PlayerFinishDiggingEvent> {
+
+    @Override
+    public void accept(PlayerFinishDiggingEvent event) {
+        event.getPlayer().removeTag(IS_DIGGING);
+    }
+}
